Rename utils to deleteAllRepoEntries and move it into the integ/repositories folder.
