+++ test/integ/handlers/on-chain-status-checker.integration.test.ts
@@ -11,7 +11,7 @@ import { log } from '../../../lib/Logging'
 import { OnChainStatusCheckerMetricNames, powertoolsMetric } from '../../../lib/Metrics'
 import { LimitOrdersRepository } from '../../../lib/repositories/limit-orders-repository'
 import { dynamoConfig, MOCK_ORDER_ENTITY, MOCK_ORDER_HASH, MOCK_SIGNATURE } from '../../test-data'
-import { deleteAllRepoEntries } from '../../unit/utils'
+import { deleteAllRepoEntries } from '../repositories/deleteAllRepoEntries'
 
 jest.mock('../../../lib/handlers/check-order-status/util', () => {
   const original = jest.requireActual('../../../lib/handlers/check-order-status/util')
