+++ test/integ/repositories/deleteAllRepoEntries.ts
@@ -1,5 +1,5 @@
-import { BaseOrdersRepository } from '../../lib/repositories/base'
-import { DYNAMO_BATCH_WRITE_MAX } from '../../lib/util/constants'
+import { BaseOrdersRepository } from '../../../lib/repositories/base'
+import { DYNAMO_BATCH_WRITE_MAX } from '../../../lib/util/constants'
 
 export async function deleteAllRepoEntries(ordersRepository: BaseOrdersRepository) {
   let orders = await ordersRepository.getOrders(DYNAMO_BATCH_WRITE_MAX, { chainId: 1 })
