+++ test/integ/repositories/limit-orders-repository.test.ts
@@ -4,7 +4,7 @@ import { ORDER_STATUS, SORT_FIELDS } from '../../../lib/entities/Order'
 import { LimitOrdersRepository } from '../../../lib/repositories/limit-orders-repository'
 import { generateRandomNonce } from '../../../lib/util/nonce'
 import { currentTimestampInSeconds } from '../../../lib/util/time'
-import { deleteAllRepoEntries } from '../../unit/utils'
+import { deleteAllRepoEntries } from './deleteAllRepoEntries'
 
 jest.mock('../../../lib/util/time')
